requirements.txt
Django>=1.8.0,<1.10.0
djangorestframework~=3.3.0
jsonfield~=1.0.0
django-environ~=0.4.0
django-autoslug~=1.9.0
django-enumfields~=0.7.0
toml~=0.9.0
django-reversion~=1.10.0
